+++ 2.2.2. spring_mvc/src/main/java/web/service/UserService.java
@@ -5,13 +5,13 @@ import web.model.User;
 import java.util.List;
 
 public interface UserService {
-    User getUserId(Integer  id);
+    User getUserId(Long  id);
 
     List<User> listUser();
 
     void saveUser(User user);
 
-    void deleteById(Integer  id);
+    void deleteById(Long  id);
 
     void updateUser(User user);
 }
